+++ srcs/execution_start.c
@@ -1,6 +1,6 @@
-# include "execute.h"
+#include "execute.h"
 
-enum stdfd_mode
+enum e_stdfd_mode
 {
 	SAVE,
 	RESTORE
@@ -42,7 +42,6 @@ int	execute_loop(t_execdata *data)
 			xclose(data->pipefd[READ]);
 			if (data->next != NULL)
 				ft_dup2(data->pipefd[WRITE], STDOUT_FILENO);
-				// data->out_fd = data->pipefd[WRITE];
 			if (setdata_cmdline_redirect(data) == 0)
 				execute_command(data);
 			exit(*(data->status));
@@ -73,6 +72,7 @@ static void	stdfd_handler(t_execdata *data, int mode)
 	}
 }
 
+//confirm WIFEXITED and WIFSIGNALED etc..
 void	execute_start(t_execdata *data)
 {
 	int			lastchild_pid;
@@ -91,8 +91,8 @@ void	execute_start(t_execdata *data)
 	{
 		lastchild_pid = execute_loop(data);
 		xwaitpid(lastchild_pid, &wstatus, 0);
-		*(data->status) = WEXITSTATUS(wstatus);//confirm WIFEXITED and WIFSIGNALED etc..
-		while(data->next)
+		*(data->status) = WEXITSTATUS(wstatus);
+		while (data->next)
 		{
 			xwaitpid(0, NULL, 0);
 			data = data->next;
