+++ srcs/wrapper2.c
@@ -17,7 +17,7 @@ char	**ft_xsplit(char *src_str, char cut_char)
 {
 	char	**splited_str;
 
-	splited_str = ft_split(src_str, cut_char); 
+	splited_str = ft_split(src_str, cut_char);
 	if (!splited_str)
 	{
 		perror("malloc");
@@ -62,35 +62,4 @@ int	xdup(int oldfd)
 		exit(EXIT_FAILURE);
 	}
 	return (newfd);
-}
-
-void	xwaitpid(pid_t pid, int *wstatus, int options)
-{
-	if (waitpid(pid, wstatus, options) == -1)
-	{
-		perror("waitpid");
-		exit(EXIT_FAILURE);
-	}
-}
-
-void	xpipe(int *pipefd)
-{
-	if (pipe(pipefd) == -1)
-	{
-		perror("pipe");
-		exit(EXIT_FAILURE);
-	}
-}
-
-pid_t	xfork(void)
-{
-	int	pid;
-
-	pid = fork();
-	if (pid == -1)
-	{
-		perror("fork");
-		exit(EXIT_FAILURE);
-	}
-	return (pid);
 }
