+++ srcs/wrapper3.c
@@ -1,0 +1,32 @@
+#include "execute.h"
+
+void	xwaitpid(pid_t pid, int *wstatus, int options)
+{
+	if (waitpid(pid, wstatus, options) == -1)
+	{
+		perror("waitpid");
+		exit(EXIT_FAILURE);
+	}
+}
+
+void	xpipe(int *pipefd)
+{
+	if (pipe(pipefd) == -1)
+	{
+		perror("pipe");
+		exit(EXIT_FAILURE);
+	}
+}
+
+pid_t	xfork(void)
+{
+	int	pid;
+
+	pid = fork();
+	if (pid == -1)
+	{
+		perror("fork");
+		exit(EXIT_FAILURE);
+	}
+	return (pid);
+}
